Break down XFormPrimView into Usd and Fabric backends
