+++ source/isaaclab/isaaclab/sim/views/xform_usd_backend.py
@@ -1,0 +1,220 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+from __future__ import annotations
+
+from collections.abc import Sequence
+
+import numpy as np
+import torch
+
+from pxr import Gf, Sdf, Usd, UsdGeom, Vt
+
+
+class UsdBackend:
+    """USD-based transform backend for :class:`XformPrimView`.
+
+    Reads and writes transforms through USD's ``xformOp`` attributes, using
+    :class:`UsdGeom.XformCache` for efficient world-space computations.
+    """
+
+    def __init__(self, prims: list[Usd.Prim], device: str):
+        self._prims = prims
+        self._device = device
+        self._ALL_INDICES = list(range(len(prims)))
+
+    @property
+    def count(self) -> int:
+        """Number of prims managed by this backend."""
+        return len(self._prims)
+
+    def initialize(self) -> None:
+        """No-op for the USD backend (no deferred setup needed)."""
+
+    # ------------------------------------------------------------------
+    # Setters
+    # ------------------------------------------------------------------
+
+    def set_world_poses(
+        self,
+        positions: torch.Tensor | None = None,
+        orientations: torch.Tensor | None = None,
+        indices: Sequence[int] | None = None,
+    ) -> None:
+        """Set world-space poses by converting to local space and writing to USD."""
+        # Resolve indices
+        if indices is None or indices == slice(None):
+            indices_list = self._ALL_INDICES
+        else:
+            indices_list = indices.tolist() if isinstance(indices, torch.Tensor) else list(indices)
+
+        # Validate inputs
+        if positions is not None:
+            if positions.shape != (len(indices_list), 3):
+                raise ValueError(
+                    f"Expected positions shape ({len(indices_list)}, 3), got {positions.shape}. "
+                    "Number of positions must match the number of prims in the view."
+                )
+            positions_array = Vt.Vec3dArray.FromNumpy(positions.cpu().numpy())
+        else:
+            positions_array = None
+        if orientations is not None:
+            if orientations.shape != (len(indices_list), 4):
+                raise ValueError(
+                    f"Expected orientations shape ({len(indices_list)}, 4), got {orientations.shape}. "
+                    "Number of orientations must match the number of prims in the view."
+                )
+            orientations_array = Vt.QuatdArray.FromNumpy(orientations.cpu().numpy())
+        else:
+            orientations_array = None
+
+        xform_cache = UsdGeom.XformCache(Usd.TimeCode.Default())
+
+        with Sdf.ChangeBlock():
+            for idx, prim_idx in enumerate(indices_list):
+                prim = self._prims[prim_idx]
+                parent_prim = prim.GetParent()
+
+                world_pos = positions_array[idx] if positions_array is not None else None
+                world_quat = orientations_array[idx] if orientations_array is not None else None
+
+                if parent_prim.IsValid() and parent_prim.GetPath() != Sdf.Path.absoluteRootPath:
+                    if positions_array is None or orientations_array is None:
+                        prim_tf = xform_cache.GetLocalToWorldTransform(prim)
+                        prim_tf.Orthonormalize()
+                        if world_pos is not None:
+                            prim_tf.SetTranslateOnly(world_pos)
+                        if world_quat is not None:
+                            prim_tf.SetRotateOnly(world_quat)
+                    else:
+                        prim_tf = Gf.Matrix4d()
+                        prim_tf.SetTranslateOnly(world_pos)
+                        prim_tf.SetRotateOnly(world_quat)
+
+                    parent_world_tf = xform_cache.GetLocalToWorldTransform(parent_prim)
+                    local_tf = prim_tf * parent_world_tf.GetInverse()
+                    local_pos = local_tf.ExtractTranslation()
+                    local_quat = local_tf.ExtractRotationQuat()
+                else:
+                    local_pos = world_pos
+                    local_quat = world_quat
+
+                if local_pos is not None:
+                    prim.GetAttribute("xformOp:translate").Set(local_pos)
+                if local_quat is not None:
+                    prim.GetAttribute("xformOp:orient").Set(local_quat)
+
+    def set_local_poses(
+        self,
+        translations: torch.Tensor | None = None,
+        orientations: torch.Tensor | None = None,
+        indices: Sequence[int] | None = None,
+    ) -> None:
+        """Set local-space poses directly on USD xformOp attributes."""
+        if indices is None or indices == slice(None):
+            indices_list = self._ALL_INDICES
+        else:
+            indices_list = indices.tolist() if isinstance(indices, torch.Tensor) else list(indices)
+
+        if translations is not None:
+            if translations.shape != (len(indices_list), 3):
+                raise ValueError(f"Expected translations shape ({len(indices_list)}, 3), got {translations.shape}.")
+            translations_array = Vt.Vec3dArray.FromNumpy(translations.cpu().numpy())
+        else:
+            translations_array = None
+        if orientations is not None:
+            if orientations.shape != (len(indices_list), 4):
+                raise ValueError(f"Expected orientations shape ({len(indices_list)}, 4), got {orientations.shape}.")
+            orientations_array = Vt.QuatdArray.FromNumpy(orientations.cpu().numpy())
+        else:
+            orientations_array = None
+
+        with Sdf.ChangeBlock():
+            for idx, prim_idx in enumerate(indices_list):
+                prim = self._prims[prim_idx]
+                if translations_array is not None:
+                    prim.GetAttribute("xformOp:translate").Set(translations_array[idx])
+                if orientations_array is not None:
+                    prim.GetAttribute("xformOp:orient").Set(orientations_array[idx])
+
+    def set_scales(self, scales: torch.Tensor, indices: Sequence[int] | None = None) -> None:
+        """Set scales on USD ``xformOp:scale`` attributes."""
+        if indices is None or indices == slice(None):
+            indices_list = self._ALL_INDICES
+        else:
+            indices_list = indices.tolist() if isinstance(indices, torch.Tensor) else list(indices)
+
+        if scales.shape != (len(indices_list), 3):
+            raise ValueError(f"Expected scales shape ({len(indices_list)}, 3), got {scales.shape}.")
+
+        scales_array = Vt.Vec3dArray.FromNumpy(scales.cpu().numpy())
+
+        with Sdf.ChangeBlock():
+            for idx, prim_idx in enumerate(indices_list):
+                prim = self._prims[prim_idx]
+                prim.GetAttribute("xformOp:scale").Set(scales_array[idx])
+
+    # ------------------------------------------------------------------
+    # Getters
+    # ------------------------------------------------------------------
+
+    def get_world_poses(self, indices: Sequence[int] | None = None) -> tuple[torch.Tensor, torch.Tensor]:
+        """Get world-space poses via :class:`UsdGeom.XformCache`."""
+        if indices is None or indices == slice(None):
+            indices_list = self._ALL_INDICES
+        else:
+            indices_list = indices.tolist() if isinstance(indices, torch.Tensor) else list(indices)
+
+        positions = Vt.Vec3dArray(len(indices_list))
+        orientations = Vt.QuatdArray(len(indices_list))
+        xform_cache = UsdGeom.XformCache(Usd.TimeCode.Default())
+
+        for idx, prim_idx in enumerate(indices_list):
+            prim = self._prims[prim_idx]
+            prim_tf = xform_cache.GetLocalToWorldTransform(prim)
+            prim_tf.Orthonormalize()
+            positions[idx] = prim_tf.ExtractTranslation()
+            orientations[idx] = prim_tf.ExtractRotationQuat()
+
+        positions = torch.tensor(np.array(positions), dtype=torch.float32, device=self._device)
+        orientations = torch.tensor(np.array(orientations), dtype=torch.float32, device=self._device)
+        return positions, orientations  # type: ignore
+
+    def get_local_poses(self, indices: Sequence[int] | None = None) -> tuple[torch.Tensor, torch.Tensor]:
+        """Get local-space poses via :class:`UsdGeom.XformCache`."""
+        if indices is None or indices == slice(None):
+            indices_list = self._ALL_INDICES
+        else:
+            indices_list = indices.tolist() if isinstance(indices, torch.Tensor) else list(indices)
+
+        translations = Vt.Vec3dArray(len(indices_list))
+        orientations = Vt.QuatdArray(len(indices_list))
+        xform_cache = UsdGeom.XformCache(Usd.TimeCode.Default())
+
+        for idx, prim_idx in enumerate(indices_list):
+            prim = self._prims[prim_idx]
+            prim_tf = xform_cache.GetLocalTransformation(prim)[0]
+            prim_tf.Orthonormalize()
+            translations[idx] = prim_tf.ExtractTranslation()
+            orientations[idx] = prim_tf.ExtractRotationQuat()
+
+        translations = torch.tensor(np.array(translations), dtype=torch.float32, device=self._device)
+        orientations = torch.tensor(np.array(orientations), dtype=torch.float32, device=self._device)
+        return translations, orientations  # type: ignore
+
+    def get_scales(self, indices: Sequence[int] | None = None) -> torch.Tensor:
+        """Get scales from USD ``xformOp:scale`` attributes."""
+        if indices is None or indices == slice(None):
+            indices_list = self._ALL_INDICES
+        else:
+            indices_list = indices.tolist() if isinstance(indices, torch.Tensor) else list(indices)
+
+        scales = Vt.Vec3dArray(len(indices_list))
+
+        for idx, prim_idx in enumerate(indices_list):
+            prim = self._prims[prim_idx]
+            scales[idx] = prim.GetAttribute("xformOp:scale").Get()
+
+        return torch.tensor(np.array(scales), dtype=torch.float32, device=self._device)
